[IMP] stock: restrict product selection to respective templates
Before commit:
--------------------------------------
- In the stock.quant view (opened by clicking the 'On Hand' value in the
  forecast report), able to update the quant of any product, even if it didn’t
  belong to the respective product/template.
- When there was no quants for product and the 'Update Quantity' button was
  clicked from the forecast report, the stock.quant no-content page appeared
  with raw HTML content.

With this commit:
--------------------------------------
- When updating a quant, only products belonging to the respective
  product/template can be selected.
- Pre-fill the product if only one variant exists.
- The stock.quant no-content page is now correctly displayed.

Task ID: [4599942](https://www.odoo.com/odoo/project/966/tasks/4599942)
