+++ addons/stock/static/src/stock_forecasted/forecasted_buttons.js
@@ -1,6 +1,6 @@
 import { _t } from "@web/core/l10n/translation";
 import { useService } from "@web/core/utils/hooks";
-import { Component } from "@odoo/owl";
+import { Component, markup } from "@odoo/owl";
 
 export class ForecastedButtons extends Component {
     static template = "stock.ForecastedButtons";
@@ -51,6 +51,9 @@ export class ForecastedButtons extends Component {
         const action = await this.orm.call(this.resModel, "action_open_quants", [[this.productId]]);
         if (action.res_model === "stock.quant") { // Quant view in inventory mode.
             action.views = [[false, "list"]];
+        }
+        if (action.help) {
+            action.help = markup(action.help);
         }
         return this.actionService.doAction(action, { onClose: this._onClose.bind(this) });
     }
