+++ addons/stock/static/src/stock_forecasted/forecasted_header.js
@@ -14,22 +14,11 @@ export class ForecastedHeader extends Component {
     }
 
     async _onClickInventory(){
-        const context = this._getActionContext();
-        const action = await this.orm.call('stock.quant', 'action_view_quants', [], { context });
+        const productIds = this.props.docs.product_variants_ids;
+        const action = await this.orm.call('product.product', 'action_open_quants', [productIds]);
         if (action.help) {
             action.help = markup(action.help);
         }
         return this.action.doAction(action);
-    }
-
-    _getActionContext(){
-        const context = { ...this.context };
-        const templates = this.props.docs.product_templates_ids;
-        if (templates) {
-            context.search_default_product_tmpl_id = templates;
-        } else {
-            context.search_default_product_id = this.props.docs.product_variants_ids;
-        }
-        return context;
     }
 }
